feat: anthropic invoke with function call
