+++ anthropic/actions/code.ts
@@ -1,6 +1,6 @@
+import { shortcircuit } from "@deco/deco";
 import { AppContext } from "../mod.ts";
 import { Anthropic } from "../deps.ts";
-import { shortcircuit } from "@deco/deco";
 export interface Props {
   /**
    * @description The system prompt to be used for the AI Assistant.
@@ -13,14 +13,7 @@ export interface Props {
   /**
    * @description The model that will complete your prompt.
    */
-  model?:
-    | "claude-3-5-sonnet-20240620"
-    | "claude-3-opus-20240229"
-    | "claude-3-sonnet-20240229"
-    | "claude-3-haiku-20240307"
-    | "claude-2.1"
-    | "claude-2.0"
-    | "claude-instant-1.2";
+  model?: Anthropic.Model;
   /**
    * @description The maximum number of tokens to generate.
    *
@@ -29,20 +22,27 @@ export interface Props {
    */
   max_tokens?: number;
 }
+
 export default async function chat(
-  { system, messages, model = "claude-3-opus-20240229", max_tokens = 4096 }:
-    Props,
+  {
+    system,
+    messages,
+    model = "claude-3-opus-20240229",
+    max_tokens = 4096,
+  }: Props,
   _req: Request,
   ctx: AppContext,
 ) {
   if (!messages) {
     return shortcircuit(new Response("No messages provided", { status: 400 }));
   }
+
   const msg = await ctx.anthropic.messages.create({
     system,
     model,
     max_tokens,
     messages,
   });
+
   return msg;
 }
